vc: updating block device support checking
The current logic needs updating, as itt assumes you want to know that
the agent supports  block devices and that block device hotplug is supported.

In reality, we just want to know that block devices are supported, since
we could either do coldplug or hotplug.

Need to update the callers of the function to make sure they explicitly
are checking for hotplug if this is the intention.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>