FC-hacking: split createContainer work between create and start
Need to be able to configure the  container without the agent being up,
and finish creating the container during OCI start flow. In docker case
we could do this sooner.  In K8S case we'll need to add concept of
sandbox state in order to know when we have received the container
workoad...

Signed-off-by: Eric Ernst <eric.ernst@intel.com>