virtcontainers: fc: Track VMM state
In order to understand when we can add devices and resources to a
firecracker VM, we need to track the VMM state.
Firecracker is a VMM that does not support hotplugging and that follows
a 2 step initialization process:

1) Launch the VMM binary itself (firecracker).
2) Send REST API commands to prepare the VM to be launched.

So keeping track of the VMM state will allow us to know when we can
send the REST API commands or if we have to queue them.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>